move config trait, remove producer late connect
