+++ src/Kafka/Conf/KafkaConfigTrait.php
@@ -1,8 +1,6 @@
 <?php
 
-namespace Jobcloud\Messaging\Kafka\Helper;
-
-use Jobcloud\Messaging\Kafka\Conf\KafkaConfiguration;
+namespace Jobcloud\Messaging\Kafka\Conf;
 
 trait KafkaConfigTrait
 {
@@ -25,7 +23,7 @@ trait KafkaConfigTrait
             $conf->set($name, is_bool($value) ? (false === $value ? 'false' : 'true') : $value);
         }
 
-        $conf->set('metadata.broker.list', implode(',', $this->kafkaConfiguration->getBrokers()));
+        $conf->set('metadata.broker.list', implode(',', $conf->getBrokers()));
 
         return $conf;
     }
