+++ src/Kafka/Consumer/KafkaConsumerBuilder.php
@@ -6,7 +6,7 @@ namespace Jobcloud\Messaging\Kafka\Consumer;
 
 use Jobcloud\Messaging\Kafka\Callback\KafkaErrorCallback;
 use Jobcloud\Messaging\Kafka\Conf\KafkaConfiguration;
-use Jobcloud\Messaging\Kafka\Helper\KafkaConfigTrait;
+use Jobcloud\Messaging\Kafka\Conf\KafkaConfigTrait;
 use RdKafka\Consumer as RdKafkaLowLevelConsumer;
 use RdKafka\KafkaConsumer as RdKafkaHighLevelConsumer;
